Refactor POS routes to extend Store API directly (mirror agentic)
The earlier wrapper-delegation pattern (POS routes hold an injected
Store API delegate) was an unnecessary layer. Agentic commerce uses a
simpler pattern that POS should mirror for consistency: each route
subclasses the Store API route directly and overrides only what is
POS-specific.

- CartAddItem now extends Store API's concrete CartAddItem, overriding
  get_args() to substitute the permission callback, and requires_nonce()
  to return false (CSRF is not a vector for non-cookie-authenticated
  POS requests; matches agentic's CheckoutSessions::requires_nonce).
- The standalone NonceCheckPolicy filter class is dropped — the
  per-route override replaces it. Two mechanisms doing the same thing
  was confusing; the per-route shape is the agentic convention.
- AbstractRoute (the wrapper base) is dropped — no longer needed.
- Routes\Controller constructs routes via SchemaController directly,
  same as Store API's own RoutesController, instead of fetching
  delegates from the Store API container.

A nice consequence: because the POS route IS now a Store API route
class registered at a different URL, AbstractCartRoute::add_response_headers
emits a Cart-Token response header automatically. Mobile clients
capture it and replay on subsequent requests via the Cart-Token
request header. Authentication::maybe_use_store_api_session_handler
detects the header and swaps in the existing StoreApi\SessionHandler
for those requests. This resolves the "session continuity across
requests" question that was previously the load-bearing follow-up;
no need to un-final StoreApi\SessionHandler.

DECISIONS.md updated to reflect the new pattern.
