+++ plugins/woocommerce/includes/class-woocommerce.php
@@ -407,7 +407,6 @@ final class WooCommerce {
 		$container->get( Automattic\WooCommerce\Internal\Orders\PointOfSaleEmailHandler::class )->register();
 		$container->get( Automattic\WooCommerce\Internal\POS\StoreApi\PolicyHooks\SessionHandlerSwap::class )->register();
 		$container->get( Automattic\WooCommerce\Internal\POS\StoreApi\PolicyHooks\StockPolicy::class )->register();
-		$container->get( Automattic\WooCommerce\Internal\POS\StoreApi\PolicyHooks\NonceCheckPolicy::class )->register();
 		$container->get( Automattic\WooCommerce\Internal\POS\StoreApi\Routes\Controller::class )->register();
 		$container->get( Automattic\WooCommerce\Internal\ShopperLists\ShopperListsController::class )->register();
